gather: moved peerMesh to a dedicated file, implemented inefficient clique finding alogirthm.
